dist.py, temp.py code consistency
Made both dist.py and temp.py more consistent with the new speed and weight code in  terms of how the value is returned as well as changing temp.py to ask what the user wants their temperature to be converted to instead of converting it to all other units.

Also moved ureg unit registry and Q_ definitions to outside the while loop(for temp.py) and inside the main function (for dist.py).

dist.py unit dictionaries now changed into just one and using numbers as an index to select the units instead (thanks copilot this is pretty cool)

Changed import statement to only import UnitRegistry from Pint

Made return type into a tuple instead of a list (don't know why I wanted a list in the first place)